+++ core/types/variant_helper.h
@@ -18,6 +18,27 @@ namespace Fresa
     template<typename T>
     inline constexpr bool is_variant_v=is_variant<T>::value;
     
+    //---Variant from tuple---
+    template <typename... Ts>
+    std::variant<Ts...> variant_from_tuple (std::tuple<Ts...>);
+    
+    //---Filter variant---
+    template <template <typename> class C, typename T>
+    std::enable_if_t<true == C<T>::value, std::tuple<>> filter_t ();
+
+    template <template <typename> class C, typename T>
+    std::enable_if_t<false == C<T>::value, std::tuple<T>> filter_t ();
+    
+    template <template <typename> class C, typename... Args>
+    struct filtered_t {
+        using value = decltype(std::tuple_cat(filter_t<C, Args>()...));
+    };
+    
+    template <template <typename> class C, typename... Args>
+    struct filter_ {
+        using value = decltype(variant_from_tuple(std::declval<typename filtered_t<C, Args...>::value>()));
+    };
+    
     //---Constexpr for---
     template<std::size_t N>
     struct num { static const constexpr auto value = N; };
@@ -52,26 +73,5 @@ namespace Fresa
     template <typename... Args0, typename... Args1>
     struct concatenate_<std::variant<Args0...>, std::variant<Args1...>> {
         using type = std::variant<Args0..., Args1...>;
-    };
-    
-    //---Variant from tuple---
-    template <typename... Ts>
-    std::variant<Ts...> variant_from_tuple (std::tuple<Ts...>);
-    
-    //---Filter variant---
-    template <template <typename> class C, typename T>
-    std::enable_if_t<true == C<T>::value, std::tuple<>> filter_t ();
-
-    template <template <typename> class C, typename T>
-    std::enable_if_t<false == C<T>::value, std::tuple<T>> filter_t ();
-    
-    template <template <typename> class C, typename... Args>
-    struct filtered_t {
-        using value = decltype(std::tuple_cat(filter_t<C, Args>()...));
-    };
-    
-    template <template <typename> class C, typename... Args>
-    struct filter_ {
-        using value = decltype(variant_from_tuple(std::declval<typename filtered_t<C, Args...>::value>()));
     };
 }