+++ serialization/reflection.h
@@ -2,11 +2,16 @@
 //by jose pazos perez
 //licensed under GPLv3 uwu
 
-//thanks to Jameson Thatcher (bluescreenofdoom)[http://bluescreenofdoom.com/post/code/Reflection/],
-//          veselink1 (refl-cpp)[https://github.com/veselink1/refl-cpp],
-//          chochlik (static reflection)[https://isocpp.org/files/papers/n3996.pdf] and
-//          KonanM (tser)[https://github.com/KonanM/tser]
-//for the help with his part
+//  Reflection in C++ is (as of now) pretty much absent, however, I wanted to implemented it to use with automatic serialization and other things.
+//  It took a lot of trial and error and many failed attempts, and it definitely has more room to grow, but I'm happy with the current result.
+//  I read many articles and source code while working on this, and I'd like to list the most useful ones here. Thank you so much!
+//      - Jameson Thatcher (bluescreenofdoom)[http://bluescreenofdoom.com/post/code/Reflection/]
+//      - KonanM (tser)[https://github.com/KonanM/tser]
+//      - veselink1 (refl-cpp)[https://github.com/veselink1/refl-cpp]
+//      - chochlik (static reflection)[https://isocpp.org/files/papers/n3996.pdf]
+//      - Alexandr Poltavsky (type loophole)[https://github.com/alexpolt/luple/blob/master/type-loophole.h]
+//      - Fabian Jung (tsmp)https://github.com/fabian-jung/tsmp
+//  In the future I'd like to keep improving this part of the project and maybe add a compiler tool to generate the reflection data.
 
 #pragma once
 
@@ -17,7 +22,11 @@
 #include <tuple>
 #include <ostream>
 #include <string_view>
+#include <numeric>
 #include "variant_helper.h"
+#include "static_str.h"
+
+#include <iostream>
 
 //---Reflection---
 namespace Fresa
@@ -79,22 +88,15 @@ namespace Fresa
         constexpr size_t str_size(char const* c, size_t str_size = 1) {
             for (; *c; ++c) ++str_size; return str_size;
         }
-        
-        //---String literals as template parameters---
-        template<size_t N>
-        struct StrL {
-            constexpr StrL(const char (&str)[N]) { std::copy_n(str, N, value); }
-            char value[N];
-        };
-        
+   
         //---Member names---
         //      struct Test : Reflection::Members<"member"> {
         //          int member;
         //      };
         //: See a macro alternative that doesn't use inheritance below
-        template<StrL... M>
+        template<Str... M>
         struct Members {
-            static constexpr std::array<const char*, sizeof...(M)> member_names { M.value... };
+            static constexpr std::array<const char*, sizeof...(M)> member_names { M.c_str()... };
         };
         
         //: Exclude certain functions from the type reflection
@@ -104,9 +106,6 @@ namespace Fresa
         };
         
         //---Struct type list---
-        //      https://github.com/apolukhin/magic_get
-        //      https://github.com/alexpolt/luple/blob/master/type-loophole.h
-        
         //: Generate friend declarations, t returns the type with auto and ct helps avoiding multiple definitions
         template<typename T, int N>
         struct tag {
@@ -153,12 +152,71 @@ namespace Fresa
         //: Get a variant type list from the type
         template<typename T, typename U>
         struct type_list;
-        template<typename T, int... NN>
-        struct type_list< T, std::integer_sequence<int, NN...> > {
-            using type = typename filter_<is_excluded, decltype(t(tag<T, NN>{}))...>::value;
+        template<typename T, int... I>
+        struct type_list< T, std::integer_sequence<int, I...> > {
+            using type = typename filter_<is_excluded, decltype(t(tag<T, I>{}))...>::value;
         };
         template<typename T>
         using as_type_list = typename type_list<T, std::make_integer_sequence<int, detect<T>(0)>>::type;
+        
+        //: Size of each member
+        template<typename U>
+        struct size_list;
+        template<typename... Ts>
+        struct size_list< std::variant<Ts...> > {
+            static constexpr std::array<size_t, sizeof...(Ts)> size{ sizeof(Ts)... };
+        };
+        template<typename T>
+        constexpr auto as_size_list = size_list<as_type_list<T>>::size;
+        
+        //: Offsets for each member
+        template<typename T, size_t index>
+        constexpr size_t get_offset() {
+            constexpr size_t s = std::accumulate(as_size_list<T>.begin(), as_size_list<T>.begin() + index, 0);
+            return s;
+        }
+        
+        //: Get member index by name (constexpr with string literal or runtime with string)
+        template<Str name, typename T, std::enable_if_t<is_reflectable<T>, bool> = true>
+        constexpr size_t get_index_c() {
+            constexpr auto it = std::find(T::member_names.begin(), T::member_names.end(), name.sv());
+            constexpr size_t index = std::distance(T::member_names.begin(), it);
+            static_assert(index != T::member_names.size(), "You tried to use a name that doesn't belong to any type");
+            return index;
+        }
+        template<typename T, std::enable_if_t<is_reflectable<T>, bool> = true>
+        size_t get_index(str name) {
+            auto it = std::find(T::member_names.begin(), T::member_names.end(), name);
+            size_t index = std::distance(T::member_names.begin(), it);
+            if (index == T::member_names.size())
+                throw std::runtime_error("[ ERROR ] You tried to use a name that doesn't belong to any type");
+            return index;
+        }
+        
+        //: Get member pointer (constexpr, by index or by name)
+        template<size_t I, typename T, std::enable_if_t<is_reflectable<T>, bool> = true>
+        constexpr auto get_member_i(T* t) {
+            constexpr size_t offset = get_offset<T, I>();
+            using M = std::variant_alternative_t<I, as_type_list<T>>;
+            return (M*)(t + offset);
+        }
+        template<Str name, typename T, std::enable_if_t<is_reflectable<T>, bool> = true>
+        constexpr auto get_member(T* t) {
+            return get_member_i<get_index_c<name, T>(), T>(t);
+        }
+        
+        //: Apply a function to a member variable by name (runtime)
+        template<typename T, typename F, std::enable_if_t<is_reflectable<T>, bool> = true>
+        auto apply(T* t, str name, F func) {
+            size_t index = get_index<T>(name);
+            for_<as_type_list<T>>([&](auto i){
+                if (i.value == index) {
+                    using M = std::variant_alternative_t<i.value, as_type_list<T>>;
+                    M* member = get_member_i<i.value, T>(t);
+                    func(member);
+                }
+            });
+        }
     }
 }
 
@@ -167,7 +225,7 @@ namespace Fresa
 //          Members(Test, member)
 //          int member;
 //      };
-//: This is the recommended way to add VertexData since it doesn't create inheritance
+//: This is the recommended way to add reflection to VertexData since it doesn't create inheritance, which makes it harder to use the default constructor
 #define Members(Type, ...) \
 static constexpr std::array<char, ::Fresa::Reflection::str_size(#__VA_ARGS__)> member_name_data = [](){ \
     std::array<char, ::Fresa::Reflection::str_size(#__VA_ARGS__)> chars{'\0'}; \